+++ lib/models/page.dart
@@ -1,0 +1,11 @@
+class Page {
+  final String name;
+  final String content;
+  final String image;
+
+  Page({
+    required this.name,
+    required this.content,
+    required this.image,
+  });
+}